+++ .storybook/preview.ts
@@ -44,14 +44,22 @@ const preview: Preview = {
 		},
 	},
 	parameters: {
-		backgrounds: { disable: true },
-		controls: {
+        backgrounds: { disable: true },
+
+        controls: {
 			matchers: {
 				color: /(background|color)$/i,
 				date: /Date$/i,
 			},
 		},
-	},
+
+        a11y: {
+            // 'todo' - show a11y violations in the test UI only
+            // 'error' - fail CI on a11y violations
+            // 'off' - skip a11y checks entirely
+            test: "todo"
+        }
+    },
 };
 
 export default preview;